# apps/taptrade-platform/go-platform/services/gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: Taya NA Predict API
  version: "1.0.0"
  description: >
    TapTrade prediction-market gateway API for non-redeemable gameplay points.
    Binary YES/NO outcomes use point-cents subunits on a 0-100 implied
    probability scale; settled winning shares return 100 point-cents per share.
    Points are gameplay-only units with no external value path.

    Coverage: this spec documents the core public, trading, portfolio,
    responsible-play, reward, loyalty, social, leaderboard, bot, auth proxy,
    admin account-review, admin discovery taxonomy, admin market operations,
    and admin market-lifecycle surfaces verified against the live handlers and
    DTOs as of 2026-06. CMS and remaining compatibility surfaces are
    intentionally out of scope here and tracked for documentation in
    IMPROVEMENT_PLAN.md P2-05 / G-04. The previous inherited spec is archived
    outside this launch document.
servers:
  - url: http://localhost:18080
    description: Local development
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: HttpOnly session cookie issued by the auth service. Mutating requests also require the X-CSRF-Token header (double-submit cookie).
    BotApiKey:
      type: http
      scheme: bearer
      description: "Bot/partner API key: Authorization: Bearer tna_<prefix>_<secret>. Scopes: read, trade."
  schemas:
...
    PredictLeaderboardEntry:
      type: object
      properties:
        boardId: { type: string }
        rank: { type: integer }
        userId: { type: string }
        displayName: { type: string }
        metricValue: { type: number }
        windowStart: { type: string, format: date-time }
        windowEnd: { type: string, format: date-time }
        settledCount: { type: integer, nullable: true }
    AdminLeaderboardBoard:
      type: object
      properties:
        leaderboardId: { type: string }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        metricKey: { type: string }
        pointMetricKey: { type: string }
        rankingMode: { type: string }
        order: { type: string }
        status: { type: string, enum: [active] }
        unit: { type: string, enum: [PTS] }
        rewardSummary: { type: string }
...
    AdminLeaderboardEventRequest:
      type: object
      properties:
        playerId: { type: string }
        score: { type: number }
        activitySourceType: { type: string }
        activitySourceId: { type: string }
        idempotencyKey: { type: string }
        metadata:
          type: object
          additionalProperties: { type: string }
        recordedAt: { type: string, format: date-time }
      required: [playerId, score, idempotencyKey]
    AdminLeaderboardEvent:
      type: object
      properties:
        eventId: { type: string }
        leaderboardId: { type: string }
        playerId: { type: string }
        score: { type: number }
        activitySourceType: { type: string }
        activitySourceId: { type: string }
        idempotencyKey: { type: string }
        metadata:
          type: object
          additionalProperties: { type: string }
        recordedAt: { type: string, format: date-time }
        unit: { type: string, enum: [PTS] }
...
paths:
...
  /api/v1/leaderboards/{id}/entries:
    get:
      summary: List public leaderboard entries
      description: Returns top ranking rows for one computed board. Optional userId is only allowed for the authenticated viewer's own rank.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: userId, in: query, schema: { type: string } }
      responses:
        "200":
          description: Leaderboard ranking rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  boardId: { type: string }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/PredictLeaderboardEntry" }
                  totalCount: { type: integer }
                  limit: { type: integer }
                  viewerEntry: { $ref: "#/components/schemas/PredictLeaderboardEntry" }
        "403": { description: Viewer-rank lookup requested for a different user }
...
  /api/v1/admin/leaderboards:
    get:
      summary: Admin list of computed prediction leaderboard boards
      description: Read-only list of code-defined computed boards with PTS unit and non-redeemable reward-summary copy.
      security: [{ SessionCookie: [] }]
      responses:
        "200":
          description: Computed leaderboard boards
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AdminLeaderboardBoard" }
...
  /api/v1/admin/leaderboards/{id}/entries:
    post:
      summary: Record an admin leaderboard activity score
      description: Records reviewed point-status leaderboard activity with activitySourceType/activitySourceId metadata and PTS response units.
      security: [{ SessionCookie: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminLeaderboardEventRequest" }
      responses:
        "200":
          description: Recorded leaderboard activity score
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminLeaderboardEvent" }
  /api/v1/admin/leaderboards/{id}/recompute:
    post:
      summary: Acknowledge computed leaderboard refresh
      description: Computed boards refresh in the background; this route returns a scheduling acknowledgement without moving points.
      security: [{ SessionCookie: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Recompute acknowledgement
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: scheduled }
                  message: { type: string }
...